fix: set time to zero from the start
The reset was not working properly since isWorkRunning was false
and the minutesDom were set to 25 minutes. To avoid further bugs and
complicating the code, timer is set to 00 : 00 and allow user to select
the timer to run.

Couple of changes were also done on the pomodoro.css and index.html.

Issue #13